Added interrupts
This file includes code for a basic interrupt detection on the button
for the rotary encoder.  When the button is pressed one LED turns on.
Currently the LED stays on, but this program could be improved to turn
the LED off after a certain amount of time or with a second button
press.